refactory directories
