+++ netcad/cli/_WIP_/cli_pull.py
@@ -21,8 +21,8 @@ from netcad.config import Environment, netcad_globals
 # Private Module Imports
 # -----------------------------------------------------------------------------
 
-from ..common_opts import opt_designs
-from .cli_netcad_main import cli
+from netcad.cli.common_opts import opt_designs
+from netcad.cli.cli_netcad_main import cli
 
 # -----------------------------------------------------------------------------
 #
