+++ netcad/cli/__init__.py
@@ -1,2 +1,10 @@
 #  Copyright (c) 2021 Jeremy Schulman
 #  GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
+
+from .netcad import script
+from .cli_netcad_main import cli
+
+from . import cli_build_checks
+from . import cli_build_configs
+from . import cli_design_list
+from . import cli_show_services
