+++ netcad/cli/clig_build.py
@@ -1,4 +1,10 @@
 #  Copyright (c) 2021 Jeremy Schulman
 #  GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
-from .execute_checks import execute_device_checks
+from netcad.cli.cli_netcad_main import cli
+
+
+@cli.group("build")
+def clig_build():
+    """build configs, tests, ..."""
+    pass
