+++ netcad/cli/clig_netcad_show.py
@@ -1,7 +1,7 @@
 #  Copyright (c) 2021 Jeremy Schulman
 #  GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
-from netcad.cli.netcad.cli_netcad_main import cli
+from netcad.cli.cli_netcad_main import cli
 
 # -----------------------------------------------------------------------------
 #
