--- netcad/cli/netcad/__init__.py
@@ -1,11 +1,0 @@
-#  Copyright (c) 2021 Jeremy Schulman
-#  GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
-
-from .netcad_main import script
-
-from . import cli_build_configs
-from . import cli_build_checks
-from . import cli_design_list
-from . import cli_pull
-from . import cli_push
-from . import cli_show_services
