--- netcad/cli/netcam/__init__.py
@@ -1,10 +1,0 @@
-#  Copyright (c) 2021 Jeremy Schulman
-#  GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
-
-from .netcam import script
-
-from . import cli_check_devices
-from .show_checks import cli_show_checks
-from . import config
-from . import cli_plugins
-from . import report
