+++ netcad/topology/cli/cli_show_interfaces.py
@@ -26,7 +26,7 @@ from netcad.logger import get_logger
 from netcad.cli import keywords
 from netcad.cli.common_opts import opt_devices, opt_designs
 from netcad.cli.device_inventory import get_devices_from_designs
-from netcad.cli.netcad.clig_netcad_show import clig_design_show
+from netcad.cli.clig_netcad_show import clig_design_show
 
 # -----------------------------------------------------------------------------
 #
