+++ netcam/cli/__init__.py
@@ -1,0 +1,5 @@
+from .netcam import script
+
+from . import config
+from . import report
+from . import show_checks
