+++ netcam/cli/cli_netcam_show.py
