+++ netcam/cli/cli_plugins.py
