+++ netcam/cli/config/cli_config_backup.py
@@ -16,11 +16,11 @@ import asyncio
 from netcad.config import netcad_globals
 from netcad.logger import get_logger
 from netcad.device import Device
-from netcad.netcam.dev_config import AsyncDeviceConfigurable
+from netcam.dev_config import AsyncDeviceConfigurable
 from netcad.cli.device_inventory import get_devices_from_designs
 from netcad.cli.common_opts import opt_devices, opt_designs, opt_configs_dir
 
-from netcad.cli.netcam.netcam_filter_devices import netcam_filter_devices
+from netcam.cli.netcam_filter_devices import netcam_filter_devices
 from .config_main import clig_config
 from .task_backup_config import backup_device_config
