+++ netcam/cli/config/cli_config_check.py
@@ -6,35 +6,29 @@ from typing import Tuple
 from pathlib import Path
 import asyncio
 
-import click
-
 from netcad.config import netcad_globals
 from netcad.logger import get_logger
 from netcad.device import Device, DeviceNonExclusive
-from netcad.netcam.dev_config import AsyncDeviceConfigurable
+from netcam.dev_config import AsyncDeviceConfigurable
 from netcad.cli.device_inventory import get_devices_from_designs
 from netcad.cli.common_opts import opt_devices, opt_designs, opt_configs_dir
-from netcad.cli.netcam.netcam_filter_devices import netcam_filter_devices
+from netcam.cli.netcam_filter_devices import netcam_filter_devices
 
 from .config_main import clig_config
-from .task_config_push import push_device_config
+from .task_config_check import check_device_config
 
 
-@clig_config.command("push")
+@clig_config.command("check")
 @opt_devices()
 @opt_designs()
 @opt_configs_dir()
-@click.option(
-    "--timeout-min",
-    help="reachability timeout (minutes)",
-    type=click.IntRange(min=1, max=5),
-    default=1,
-)
 def cli_netcam_config_backup(
-    devices: Tuple[str], designs: Tuple[str], configs_dir: Path, timeout_min: int
+    devices: Tuple[str],
+    designs: Tuple[str],
+    configs_dir: Path,
 ):
     """
-    Deploy the design build configurations to device(s)
+    Given the built configuration, check that it will load and save the diff.
     """
     log = get_logger()
 
@@ -43,23 +37,16 @@ def cli_netcam_config_backup(
         return
 
     use_device_objs = netcam_filter_devices(device_objs)
-    asyncio.run(
-        run_deploy_configs(
-            configs_dir=configs_dir,
-            device_objs=use_device_objs,
-            rollback_timeout=timeout_min,
-        )
-    )
+    asyncio.run(run_check_configs(configs_dir=configs_dir, device_objs=use_device_objs))
 
 
-async def run_deploy_configs(
-    device_objs: list[Device], configs_dir: Path, rollback_timeout: int
-):
+async def run_check_configs(device_objs: list[Device], configs_dir: Path):
     log = get_logger()
 
     netcam_plugins = netcad_globals.g_netcam_plugins_os_catalog
 
     dev_cfg: AsyncDeviceConfigurable
+    tasks = list()
 
     for dev_obj in device_objs:
         if not (pg_obj := netcam_plugins.get(dev_obj.os_name)):
@@ -74,16 +61,23 @@ async def run_deploy_configs(
             )
             continue
 
+        if dev_cfg.capabilities == dev_cfg.Capabilities.none:
+            log.warning(
+                f"{dev_obj.name}: Does not support configuration management, skipping."
+            )
+            continue
+
         dev_cfg.config_file = (
             configs_dir / dev_obj.design.name / (dev_obj.name + ".cfg")
         )
-        dev_cfg.config_id = f"{dev_cfg.device.name}-{os.getpid()}"
+        dev_cfg.config_id = f"{dev_cfg.device.name}-{os.getpid()}-check"
 
         # TODO: for now, we are usin the fact that the device in the design is
         #       either exclusive or non-exclusive to determine whether or not
         #       to check the config with replacing or merging the built config.
 
         dev_cfg.replace = not isinstance(dev_obj, DeviceNonExclusive)
+        tasks.append(asyncio.create_task(check_device_config(dev_cfg)))
 
-        # TODO: need to check for exceptions
-        await push_device_config(dev_cfg, rollback_timeout=rollback_timeout)
+    # TODO: need to check for excpeitons
+    await asyncio.gather(*tasks)
