+++ netcam/cli/netcam_filter_devices.py
