+++ netcam/cli/show_checks/__init__.py
@@ -1,0 +1,1 @@
+from . import cli_show_checks
