+++ netcam/cli/show_checks/cli_show_checks.py
