+++ netcam/cli/show_checks/show_design_summary.py
