+++ netcam/cli/show_checks/show_device_brief_table.py
