+++ netcam/cli/show_checks/show_device_check_logs.py
