+++ netcam/dev_config.py
