+++ pyproject.toml
@@ -1,9 +1,13 @@
 [tool.poetry]
-name = "netcad"
-version = "0.4.0"
-description = "NetCAD/CAM - Network Automation by Design"
-authors = ["Jeremy Schulman"]
-readme = "README.md"
+    name = "netcad"
+    version = "0.5.0"
+    description = "NetCAD/CAM - Network Automation by Design"
+    authors = ["Jeremy Schulman"]
+    readme = "README.md"
+    packages = [
+        { include = "netcad" },
+        { include = "netcam" },
+    ]
 
 [tool.poetry.dependencies]
 python = "^3.10"
@@ -44,5 +48,5 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-    netcad = 'netcad.cli.netcad:script'
-    netcam = 'netcad.cli.netcam:script'
+    netcad = 'netcad.cli:script'
+    netcam = 'netcam.cli:script'
